ili9488.asl: add label DC/RS & reset pins used
These pins are unconfirmed. As of commit the only thing
confirmed is that the ACPI table compiles and can
be executed with no errors via acpiexec.

May need to define a secondary ACPI table for the
MEC1705. Haven't been able to locate the named
object associated with SPI controller in th EC
nor have I found the named object associated
with the GPI controller in the EC.

Signed-off-by: Vincent Davis Jr <vince@underview.tech>